Update tool to use encoder logic
